Handle event callback differently
The Zendesk's client used for requests introduces differences compared
to the jQuery's API. The request callbacks are not called with all of
the arguments jQuery states they do, but only with `data` in case of
success, and `jqXHR` in case of failure. Thus, we need to pass any extra
information -callbacks may need- through local variables.