Use a class wrapper for CSS rules
Until now, Zendesk wrapper all application-specific CSS rules under a
class that was used to represent the application. The newer version of
apps introduces an `iframe` from applications and no such class is used,
resulting to a lower specificity for each CSS rule. In order to preserve
the behavior, we used a class for the root app element and nested all
our CSS rules under that.