Improve articles upsert operation
Currently we push resources to Transifex all at once, which causes
issues if the resources are too many or if the articles themselves contain
too much content. In order to solve this, we push resources one at a
time, uploading the next one only when the previous one has been
completed.