Pull translations from Transifex using queues
Instead of firing up a set of requests in parallel, we will now use a
queue to pull translations from Transifex, exactly the same we do when
pushing content. This way, we will no longer abuse the network's
bandwidth and we will provide an improved UX by informing the user about
the syncing progress.