Article translations with retries
The problem is that when updating article translations
multiple requests to zendesk api are
created in parallel and at some point
the zendesk service could respond with
error 503 or similar (too busy)
The solution given is to apply a retries mechanism
in order to call the failed request at some other point in time which
could lead to a successful response.
The mechanism is a backoff retry with
jitter. The number of retries is 5 per request.