Fix error/warning states in articles
When the user pushes the articles in Tx, the status for them is
depicted in the UI for each article. However, this status is not
depicted when the user gets translations from Tx. Now, when the user
gets a warning/error message on top of the list, he can also see which
articles have issues and in which locale in case of warning. The locales
are marked with either warning/error color.